refactor: #961 기존 SSE 기반 코드를 Web Socket 기반으로 변경
